+++ Homework-11-webflux/src/main/java/ru/otus/hw/services/GenreServiceImpl.java
@@ -1,10 +1,13 @@
 package ru.otus.hw.services;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 import ru.otus.hw.models.Genre;
+import ru.otus.hw.repositories.BookRepository;
 import ru.otus.hw.repositories.GenreRepository;
 
 @RequiredArgsConstructor
@@ -12,6 +15,8 @@ import ru.otus.hw.repositories.GenreRepository;
 public class GenreServiceImpl implements GenreService {
 
     private final GenreRepository genreRepository;
+
+    private final BookRepository bookRepository;
 
     @Override
     public Flux<Genre> findAll() {
@@ -30,6 +35,13 @@ public class GenreServiceImpl implements GenreService {
 
     @Override
     public Mono<Void> deleteById(String id) {
-        return genreRepository.deleteById(id);
+        return bookRepository.countByGenreId(id)
+                .flatMap(bookCount -> {
+                    if (bookCount > 0) {
+                        return Mono.error(new ResponseStatusException(HttpStatus.CONFLICT,
+                                "Cannot delete genre with associated books, kindly delete books first."));
+                    }
+                    return genreRepository.deleteById(id);
+                });
     }
 }
